requirements.txt
boto3
openai
fastapi
uvicorn
python-dotenv